automation_bot/backend/requirements.txt
python-telegram-bot==20.8
playwright==1.41.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
# Tkinter já vem com Python, não precisa instalar
